Update save-recording script to log more clear post instructions
